+++ tests/testthat/test-to_xml.R
@@ -23,6 +23,20 @@ test_that("to_xml works for Rmd", {
 })
 
 
+test_that("to_xml can parse markdown with special control characters", {
+  # skip if we are on windows with R version lower than 4.2.0
+  os <- tolower(Sys.info())[["sysname"]]
+  no_utf8_support <- os == "windows" && getRversion() < numeric_version('4.2.0')
+  skip_if(no_utf8_support, message = "this system cannot test UTF-8 output")
+
+  tmp <- withr::local_tempfile()
+  writeLines("\u2018test single\u2019 \u001C\u201Ctest double\u201D", tmp)
+  expect_no_error(xml <- tinkr::to_xml(tmp))
+  expect_equal(xml2::xml_text(xml$body), "'test single' \"test double\"")
+})
+
+
+
 test_that("to_xml will not convert numeric options to character", {
   txt <- "```{r txt, fig.width=4.2, fig.height=4.2, out.width='100%', purl = TRUE}\n#code\n```"
   con <- textConnection(txt)